Add files via upload
In this project, the dataset is from a real-world dataset from a telecom company that includes
customer information, such as demographics, usage patterns, and whether they churned (i.e., left
the company) or not. The goal was to prepare the data for analysis, explore key patterns and trends,
and create visualizations to gain insights into factors influencing customer churn.